requirements.txt
xpander-sdk
xpander-utils
openai
python-dotenv